requirements.txt
Pillow==8.0.1
stockfish==3.10.2
